_backend/requirements.txt
uvicorn
psycopg2-binary
httpx
fastapi
pydantic
PyJWT